type: refactor subject: merge interfaces and implementation\n\n Moved majority of interfaces to the implementation however two provided some difficulty IMutator due to it having multiple implementations and IReadOnlyInputComponent due to the delegate void Display would need a new home
