--- src/Stryker.Core/Stryker.Core/Compiling/IRollbackProcess.cs
@@ -1,11 +1,0 @@
-﻿using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp;
-using System.Collections.Immutable;
-
-namespace Stryker.Core.Compiling
-{
-    public interface IRollbackProcess
-    {
-        RollbackProcessResult Start(CSharpCompilation compiler, ImmutableArray<Diagnostic> diagnostics);
-    }
-}
