--- src/Stryker.Core/Stryker.Core/IStrykerRunner.cs
@@ -1,9 +1,0 @@
-﻿using Stryker.Core.Options;
-
-namespace Stryker.Core
-{
-    public interface IStrykerRunner
-    {
-        void RunMutationTest(StrykerOptions options);
-    }
-}
