--- src/Stryker.Core/Stryker.Core/Initialisation/IAssemblyReferenceResolver.cs
@@ -1,12 +1,0 @@
-﻿using Microsoft.CodeAnalysis;
-using System.Collections.Generic;
-
-namespace Stryker.Core.Initialisation
-{
-    public interface IAssemblyReferenceResolver
-    {
-        IEnumerable<PortableExecutableReference> ResolveReferences(string projectPath, string projectFileName, string projectUnderTestAssemblyName);
-        IEnumerable<string> GetAssemblyPathsFromOutput(string paths);
-        IEnumerable<string> GetReferencePathsFromOutput(IEnumerable<string> paths);
-    }
-}
