--- src/Stryker.Core/Stryker.Core/Initialisation/IINitialBuildProcess.cs
@@ -1,7 +1,0 @@
-﻿namespace Stryker.Core.Initialisation
-{
-    public interface IInitialBuildProcess
-    {
-        void InitialBuild(string path, string projectName);
-    }
-}
