--- src/Stryker.Core/Stryker.Core/Initialisation/IInitialTestProcess.cs
@@ -1,9 +1,0 @@
-﻿using Stryker.Core.TestRunners;
-
-namespace Stryker.Core.Initialisation
-{
-    public interface IInitialTestProcess
-    {
-        int InitialTest(ITestRunner testRunner);
-    }
-}
