--- src/Stryker.Core/Stryker.Core/Initialisation/IInitialisationProcess.cs
@@ -1,10 +1,0 @@
-﻿using Stryker.Core.MutationTest;
-using Stryker.Core.Options;
-
-namespace Stryker.Core.Initialisation
-{
-    public interface IInitialisationProcess
-    {
-        MutationTestInput Initialize(StrykerOptions options);
-    }
-}
