--- src/Stryker.Core/Stryker.Core/Initialisation/IInputFileResolver.cs
@@ -1,9 +1,0 @@
-﻿using Stryker.Core.MutationTest;
-
-namespace Stryker.Core.Initialisation
-{
-    public interface IInputFileResolver
-    {
-        ProjectInfo ResolveInput(string currentDirectory, string projectUnderTestNameFilter);
-    }
-}
