+++ src/Stryker.Core/Stryker.Core/Initialisation/InputFileResolver.cs
@@ -9,6 +9,11 @@ using System.Xml.Linq;
 
 namespace Stryker.Core.Initialisation
 {
+    public interface IInputFileResolver
+    {
+        ProjectInfo ResolveInput(string currentDirectory, string projectUnderTestNameFilter);
+    }
+
     /// <summary>
     ///  - Reads .csproj to find project under test
     ///  - Scans project under test and store files to mutate
