--- src/Stryker.Core/Stryker.Core/Mutants/IMutantOrchestrator.cs
@@ -1,15 +1,0 @@
-﻿using Microsoft.CodeAnalysis;
-using System.Collections.Generic;
-
-namespace Stryker.Core.Mutants
-{
-    public interface IMutantOrchestrator
-    {
-        SyntaxNode Mutate(SyntaxNode rootNode);
-        /// <summary>
-        /// Gets the stored mutants and resets the mutant list to an empty collection
-        /// </summary>
-        /// <returns>Mutants</returns>
-        IEnumerable<Mutant> GetLatestMutantBatch();
-    }
-}
