+++ src/Stryker.Core/Stryker.Core/Mutants/Mutant.cs
@@ -1,6 +1,16 @@
 ﻿namespace Stryker.Core.Mutants
 {
     /// <summary>
+    /// This interface should only contain readonly properties to ensure that others than the mutationtest proces cannot modify mutants.
+    /// </summary>
+    public interface IReadOnlyMutant
+    {
+        int Id { get; }
+        Mutation Mutation { get; }
+        MutantStatus ResultStatus { get; }
+    }
+    
+    /// <summary>
     /// Represents a single mutation on domain level
     /// </summary>
     public class Mutant : IReadOnlyMutant
