--- src/Stryker.Core/Stryker.Core/MutationTest/IMutationTestExecutor.cs
@@ -1,12 +1,0 @@
-﻿using Stryker.Core.Mutants;
-
-namespace Stryker.Core.MutationTest
-{
-    /// <summary>
-    /// Executes exactly one mutationtest and stores the result
-    /// </summary>
-    public interface IMutationTestExecutor
-    {
-        void Test(Mutant mutant);
-    }
-}
