--- src/Stryker.Core/Stryker.Core/MutationTest/IMutationTestProcess.cs
@@ -1,8 +1,0 @@
-﻿namespace Stryker.Core.MutationTest
-{
-    public interface IMutationTestProcess
-    {
-        void Mutate();
-        void Test();
-    }
-}
