+++ src/Stryker.Core/Stryker.Core/TestRunners/DotnetTestRunner.cs
@@ -4,6 +4,12 @@ using System.Collections.ObjectModel;
 
 namespace Stryker.Core.TestRunners
 {
+    public interface ITestRunner
+    {
+        TestRunResult RunAll(int? timeoutMS);
+        void SetActiveMutation(int? id);
+    }
+    
     public class DotnetTestRunner : ITestRunner
     {
         private string _path { get; set; }
