--- src/Stryker.Core/Stryker.Core/TestRunners/ITestRunner.cs
@@ -1,8 +1,0 @@
-﻿namespace Stryker.Core.TestRunners
-{
-    public interface ITestRunner
-    {
-        TestRunResult RunAll(int? timeoutMS);
-        void SetActiveMutation(int? id);
-    }
-}
