+++ src/Stryker.Core/Stryker.Core/Testing/Chalk.cs
@@ -2,6 +2,15 @@
 
 namespace Stryker.Core.Testing
 {
+    public interface IChalk
+    {
+        void Red(string text);
+        void Yellow(string text);
+        void Green(string text);
+        void DarkGray(string text);
+        void Default(string text);
+    }
+    
     public class Chalk : IChalk
     {
         public void Red(string text)
