--- src/Stryker.Core/Stryker.Core/Testing/IMetadataReferenceProxy.cs
@@ -1,9 +1,0 @@
-﻿using Microsoft.CodeAnalysis;
-
-namespace Stryker.Core.Testing
-{
-    public interface IMetadataReferenceProvider
-    {
-        PortableExecutableReference CreateFromFile(string path);
-    }
-}
