+++ src/Stryker.Core/Stryker.Core/Testing/MetadataReferenceProxy.cs
@@ -3,6 +3,11 @@ using System.Diagnostics.CodeAnalysis;
 
 namespace Stryker.Core.Testing
 {
+    public interface IMetadataReferenceProvider
+    {
+        PortableExecutableReference CreateFromFile(string path);
+    }
+    
     [ExcludeFromCodeCoverage]
     public class MetadataReferenceProvider : IMetadataReferenceProvider
     {
